+++ NEWS.md
@@ -3,6 +3,7 @@
 - remove the factor attribute of the treatment_description variable
 - add pilevel in the output data set for prediction interval level
 - replace treatment_label with treatment_description in observed data for enrollment prediction
+- update the upper bound of the cutoff reference line in prediction plot
 
 # eventPred 0.2.0